Objective function
Changed the MPC cost formulation so that, when using a delta control scheme, the controller penalizes delta_u (the change in control) instead of u = u_prev + delta_u. This avoids including the free parameter u_prev in the CasADi objective function, eliminating the "free variables" error during function initialization (this was happening when trying to optimise "R", for example).